+++ plugin/plugin/src/test/kotlin/net/twisterrob/gradle/java/JavaPluginIntgTest.kt
@@ -12,7 +12,9 @@ import org.intellij.lang.annotations.Language
 import org.junit.Test
 
 /**
+ * @see BaseJavaPlugin
  * @see JavaPlugin
+ * @see JavaLibPlugin
  */
 class JavaPluginIntgTest : BaseAndroidIntgTest() {
 
@@ -46,6 +48,102 @@ class JavaPluginIntgTest : BaseAndroidIntgTest() {
 		result.assertSuccess(":jar")
 		result.assertHasOutputLine("""Java: true""".toRegex())
 		result.assertHasOutputLine("""Java Library: true""".toRegex())
+	}
+
+	@Test fun `default-enabled warnings can be turned off`() {
+		@Language("java")
+		val uncheckedWarning = """
+			package ${packageName};
+			import java.util.ArrayList;
+			import java.util.List;
+			public class Unchecked {
+				@SuppressWarnings("UnnecessaryLocalVariable")
+				public void unchecked() {
+					List<String> strings = new ArrayList<>();
+					List<?> list = strings;
+					//noinspection unchecked suppressed in code, but javac still picks it up
+					List<Number> numbers = (List<Number>) list;
+				}
+			}
+		""".trimIndent()
+		gradle.file(uncheckedWarning, "src/main/java/${packageFolder}/Unchecked.java")
+		@Language("gradle")
+		val script = """
+			apply plugin: 'net.twisterrob.java'
+			tasks.named("compileJava").configure {
+				it.options.compilerArgs += [
+					// Disable default-enabled warnings (restore javac default state)
+					"-Xlint:-unchecked"
+				]
+			}
+		""".trimIndent()
+
+		val result = gradle.run(script, "build").build()
+
+		result.assertSuccess(":compileJava")
+		result.assertHasOutputLine("""Note: .*Unchecked.java uses unchecked or unsafe operations.""".toRegex())
+		result.assertHasOutputLine("""Note: Recompile with -Xlint:unchecked for details.""".toRegex())
+	}
+
+
+	@Test fun `unchecked warnings show up`() {
+		@Language("java")
+		val uncheckedWarning = """
+			package ${packageName};
+			import java.util.ArrayList;
+			import java.util.List;
+			public class Unchecked {
+				@SuppressWarnings("UnnecessaryLocalVariable")
+				public void unchecked() {
+					List<String> strings = new ArrayList<>();
+					List<?> list = strings;
+					//noinspection unchecked suppressed in code, but javac still picks it up
+					List<Number> numbers = (List<Number>) list;
+				}
+			}
+		""".trimIndent()
+		gradle.file(uncheckedWarning, "src/main/java/${packageFolder}/Unchecked.java")
+		@Language("gradle")
+		val script = """
+			apply plugin: 'net.twisterrob.java'
+		""".trimIndent()
+
+		val result = gradle.run(script, "build").build()
+
+		result.assertSuccess(":compileJava")
+		result.assertNoOutputLine("""Note: .*Unchecked.java uses unchecked or unsafe operations.""".toRegex())
+		result.assertNoOutputLine("""Note: Recompile with -Xlint:unchecked for details.""".toRegex())
+		result.assertHasOutputLine(""".*Unchecked.java:10: warning: \[unchecked\] unchecked cast""".toRegex())
+	}
+
+	@Test fun `deprecation warnings show up`() {
+		@Language("java")
+		val uncheckedWarning = """
+			package ${packageName};
+			public class Deprecation {
+				public void deprecation() {
+					//noinspection deprecation suppressed in code, but javac still picks it up
+					new DeprecatedClass();
+				}
+			}
+			
+			@SuppressWarnings("DeprecatedIsStillUsed")
+			@Deprecated
+			class DeprecatedClass {
+			}
+		""".trimIndent()
+		gradle.file(uncheckedWarning, "src/main/java/${packageFolder}/Deprecation.java")
+		@Language("gradle")
+		val script = """
+			apply plugin: 'net.twisterrob.java'
+		""".trimIndent()
+
+		val result = gradle.run(script, "build").build()
+
+		result.assertSuccess(":compileJava")
+		result.assertNoOutputLine("""Note: .*Deprecation.java uses or overrides a deprecated API.""".toRegex())
+		result.assertNoOutputLine("""Note: Recompile with -Xlint:deprecation for details.""".toRegex())
+		result.assertHasOutputLine(""".*Deprecation.java:5: warning: \[deprecation\] DeprecatedClass in ${packageName} has been deprecated""".toRegex())
 	}
 
 	@Test fun `test code supports Java 8`() {